1.增删改之前都需要到数据库中判断该用户是否有增删改的权限 2.查之前也需要判断是否有权限，但只要判断缓存中的权限列表就行了 3.controller中加上权限控制 4.页面中也加入权限控制 --- head.jsp 5.页面中的权限控制     -- 删除、编辑等按钮

+++ src/com/fuwei/controller/CompanyController.java
@@ -7,6 +7,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,6 +18,7 @@ import com.fuwei.commons.SystemCache;
 import com.fuwei.commons.SystemContextUtils;
 import com.fuwei.entity.Company;
 import com.fuwei.entity.User;
+import com.fuwei.service.AuthorityService;
 import com.fuwei.service.CompanyService;
 import com.fuwei.util.DateTool;
 import com.fuwei.util.HanyuPinyinUtil;
@@ -27,6 +29,8 @@ public class CompanyController extends BaseController {
 	
 	@Autowired
 	CompanyService companyService;
+	@Autowired
+	AuthorityService authorityService;
 	
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
 	@ResponseBody
@@ -34,6 +38,11 @@ public class CompanyController extends BaseController {
 			HttpServletResponse response) throws Exception{
 		
 		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "company/add";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有添加公司的权限", null);
+		}
 		company.setHelp_code(HanyuPinyinUtil.getFirstSpellByString(company.getFullname())) ;
 		company.setCreated_at(DateTool.now());
 		company.setUpdated_at(DateTool.now());
@@ -49,8 +58,14 @@ public class CompanyController extends BaseController {
 	
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.POST)
 	@ResponseBody
-	public Map<String,Object> delete(@PathVariable int id, HttpServletRequest request,
+	public Map<String,Object> delete(@PathVariable int id,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "company/delete";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有删除公司的权限", null);
+		}
 		int success = companyService.remove(id);
 		
 		//更新缓存
@@ -62,17 +77,27 @@ public class CompanyController extends BaseController {
 	
 	@RequestMapping(value = "/get/{id}", method = RequestMethod.GET)
 	@ResponseBody
-	public Company get(@PathVariable int id, HttpServletRequest request,
+	public Company get(@PathVariable int id, HttpSession session,HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		String lcode = "company/index";
+		Boolean hasAuthority = SystemCache.hasAuthority(session, lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有查看公司列表的权限", null);
+		}
 		Company company = companyService.get(id);
 		return company;
-		
 	}
 	
 	@RequestMapping(value = "/put", method = RequestMethod.POST)
 	@ResponseBody
-	public Map<String,Object> update(Company company, HttpServletRequest request,
+	public Map<String,Object> update(Company company,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "company/edit";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有编辑公司的权限", null);
+		}
 		company.setHelp_code(HanyuPinyinUtil.getFirstSpellByString(company.getFullname())) ;
 		company.setUpdated_at(DateTool.now());
 		int success = companyService.update(company);
