+++ src/com/fuwei/controller/GongXuController.java
@@ -7,6 +7,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,6 +20,7 @@ import com.fuwei.entity.Company;
 import com.fuwei.entity.GongXu;
 import com.fuwei.entity.Salesman;
 import com.fuwei.entity.User;
+import com.fuwei.service.AuthorityService;
 import com.fuwei.service.CompanyService;
 import com.fuwei.service.GongXuService;
 import com.fuwei.service.SalesmanService;
@@ -31,12 +33,20 @@ public class GongXuController extends BaseController {
 	
 	@Autowired
 	GongXuService gongxuService;
+	@Autowired
+	AuthorityService authorityService;
 	
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
 	@ResponseBody
 	public Map<String,Object> add(GongXu gongxu,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
 		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "gongxu/add";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有添加工序的权限", null);
+		}
+		
 		gongxu.setCreated_at(DateTool.now());
 		gongxu.setUpdated_at(DateTool.now());
 		gongxu.setCreated_user(user.getId());
@@ -51,10 +61,15 @@ public class GongXuController extends BaseController {
 	
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.POST)
 	@ResponseBody
-	public Map<String,Object> delete(@PathVariable int id, HttpServletRequest request,
+	public Map<String,Object> delete(@PathVariable int id, HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "gongxu/delete";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有删除工序的权限", null);
+		}
 		int success = gongxuService.remove(id);
-		
 		//更新缓存
 		new SystemCache().initGongxuList();
 		
@@ -64,8 +79,13 @@ public class GongXuController extends BaseController {
 	
 	@RequestMapping(value = "/get/{id}", method = RequestMethod.GET)
 	@ResponseBody
-	public GongXu get(@PathVariable int id, HttpServletRequest request,
+	public GongXu get(@PathVariable int id,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		String lcode = "gongxu/index";
+		Boolean hasAuthority = SystemCache.hasAuthority(session, lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有查看公司列表的权限", null);
+		}
 		GongXu gongxu = gongxuService.get(id);
 		return gongxu;
 		
@@ -73,8 +93,14 @@ public class GongXuController extends BaseController {
 	
 	@RequestMapping(value = "/put", method = RequestMethod.POST)
 	@ResponseBody
-	public Map<String,Object> update(GongXu gongxu, HttpServletRequest request,
+	public Map<String,Object> update(GongXu gongxu,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "gongxu/edit";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有编辑工序的权限", null);
+		}
 		gongxu.setUpdated_at(DateTool.now());
 		int success = gongxuService.update(gongxu);
