+++ src/com/fuwei/controller/QuoteController.java
@@ -10,6 +10,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -23,6 +24,7 @@ import com.fuwei.entity.Quote;
 import com.fuwei.entity.QuotePrice;
 import com.fuwei.entity.Sample;
 import com.fuwei.entity.User;
+import com.fuwei.service.AuthorityService;
 import com.fuwei.service.QuotePriceService;
 import com.fuwei.service.QuoteService;
 import com.fuwei.util.DateTool;
@@ -35,12 +37,19 @@ public class QuoteController extends BaseController{
 	QuoteService quoteService;
 	@Autowired
 	QuotePriceService quotePriceService;
+	@Autowired
+	AuthorityService authorityService;
 	
 	//获取未确认提交报价列表
 	@RequestMapping(value = "/index", method = RequestMethod.GET)
 	@ResponseBody
 	public ModelAndView index(HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		String lcode = "quote/index";
+		Boolean hasAuthority = SystemCache.hasAuthority(session, lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有查看报价列表的权限", null);
+		}
 		List<Quote> quotelist = quoteService.getDetailList();
 		HashMap<Integer, List<Quote>> quoteMap=new HashMap<Integer, List<Quote>>();
 		for (Quote quote : quotelist) {
@@ -64,6 +73,13 @@ public class QuoteController extends BaseController{
 			HttpServletResponse response) throws Exception{
 		
 		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "quote/add";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有添加报价的权限", null);
+		}
+		
+		
 		QuotePrice quoteprice = quotePriceService.get(quotePriceId);
 		if(quoteprice == null){
 			throw new Exception("没有相关的公司价格");
@@ -81,8 +97,16 @@ public class QuoteController extends BaseController{
 	
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.POST)
 	@ResponseBody
-	public Map<String,Object> delete(@PathVariable int id, HttpServletRequest request,
+	public Map<String,Object> delete(@PathVariable int id,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		
+		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "quote/delete";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有删除报价的权限", null);
+		}
+		
 		int success = quoteService.remove(id);
 		return this.returnSuccess();
