+++ src/com/fuwei/controller/SalesmanController.java
@@ -7,6 +7,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.PermissionDeniedDataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,6 +19,7 @@ import com.fuwei.commons.SystemContextUtils;
 import com.fuwei.entity.Company;
 import com.fuwei.entity.Salesman;
 import com.fuwei.entity.User;
+import com.fuwei.service.AuthorityService;
 import com.fuwei.service.CompanyService;
 import com.fuwei.service.SalesmanService;
 import com.fuwei.util.DateTool;
@@ -29,12 +31,21 @@ public class SalesmanController extends BaseController {
 	
 	@Autowired
 	SalesmanService salesmanService;
+	@Autowired
+	AuthorityService authorityService;
 	
 	@RequestMapping(value = "/add", method = RequestMethod.POST)
 	@ResponseBody
 	public Map<String,Object> add(Salesman salesman,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		
 		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "salesman/add";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有添加业务员的权限", null);
+		}
+		
 		salesman.setHelp_code(HanyuPinyinUtil.getFirstSpellByString(salesman.getName())) ;
 		salesman.setCreated_at(DateTool.now());
 		salesman.setUpdated_at(DateTool.now());
@@ -50,8 +61,16 @@ public class SalesmanController extends BaseController {
 	
 	@RequestMapping(value = "/delete/{id}", method = RequestMethod.POST)
 	@ResponseBody
-	public Map<String,Object> delete(@PathVariable int id, HttpServletRequest request,
+	public Map<String,Object> delete(@PathVariable int id,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		
+		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "salesman/delete";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有删除业务员的权限", null);
+		}
+		
 		int success = salesmanService.remove(id);
 		
 		//更新缓存
@@ -63,8 +82,13 @@ public class SalesmanController extends BaseController {
 	
 	@RequestMapping(value = "/get/{id}", method = RequestMethod.GET)
 	@ResponseBody
-	public Salesman get(@PathVariable int id, HttpServletRequest request,
+	public Salesman get(@PathVariable int id,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		String lcode = "salesman/index";
+		Boolean hasAuthority = SystemCache.hasAuthority(session, lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有查看业务员列表的权限", null);
+		}
 		Salesman salesman = salesmanService.get(id);
 		return salesman;
 		
@@ -72,8 +96,16 @@ public class SalesmanController extends BaseController {
 	
 	@RequestMapping(value = "/put", method = RequestMethod.POST)
 	@ResponseBody
-	public Map<String,Object> update(Salesman salesman, HttpServletRequest request,
+	public Map<String,Object> update(Salesman salesman,HttpSession session, HttpServletRequest request,
 			HttpServletResponse response) throws Exception{
+		
+		User user = SystemContextUtils.getCurrentUser(session).getLoginedUser();
+		String lcode = "salesman/edit";
+		Boolean hasAuthority = authorityService.checkLcode(user.getId(), lcode);
+		if(!hasAuthority){
+			throw new PermissionDeniedDataAccessException("没有编辑业务员的权限", null);
+		}
+		
 		salesman.setHelp_code(HanyuPinyinUtil.getFirstSpellByString(salesman.getName())) ;
 		salesman.setUpdated_at(DateTool.now());
 		int success = salesmanService.update(salesman);
